requirements.txt
weaviate-client==3.24.2
torch==2.0.1
transformers
youtube-transcript-api==0.6.1
pytube==15.0.0
langchain==0.0.319
streamlit==1.27.2
openai==0.28.1
pandas==2.1.1
python-dotenv==1.0.0
google-api-python-client

extra-streamlit-components
# pip install extra-streamlit-components